Rework some unit test patterns
In order to check the invokation order of various hooks, the tests
were previously implemented as a sort of running tabulation. This is
incredibly hard to debug when a test fails. Instead, I modified the code
to simply append descriptive strings to a slice. This allows us to easily
track which method is being called and in which order.

Signed-off-by: John Schnake <jschnake@vmware.com>