define cluster provider interface and migrate kind and kwok
This implementation defines standard interface for cluster providers
that can be used with the e2e-framework. Currently, framework only had
support for kind as a provider. With new requirements coming around the
integration needs for providers such as `kwok` or `k3d` defining a
standard interface becomes an essential need for the framework.

Along with defining the Provider interface, this PR also make a few
changes on the following lines.

1. Cleanup unwnated internal/testutils package that is no longer
   required
2. Update golangci-lint dependencies
3. Fix multi cluster test cases that were flaky
4. Add ability to log the STDOUT of `kind create cluster` command on
   failures
5. fix helm custom binary lookup logic
6. migrate kwok provider to use the new interface