Add condition for checking if a deployment is available
This change adds a new condition to check if a deployment is available.
The wait condition only requires the user to provide the deployment name
and namespace and the condition method handles building the deployment
object/checking the deployment condition for a match.

With this new condition, it will remove the need for the user to build
the deployment object and call deployment condition match condition in
their code bases.