[env] Don't run next assess when failNow is called
When writing tests, this allows the writer to control whether the
test should continue running the next assess when a failure occurs.
Basically, they can chose between using `t.FailNow()` or `t.Fail()`
depending on if the failure should stop the test or not.
It works independently from the `-fail-fast` flag specific to the
framework and allows a more granular control on test failures.