[env] Do not share config between tests
This commit fixes a bug where the config object was shared between tests.
When running tests in parallel, it was then impossible to rely on fields
like the namespace because they could have been overwritten by another
test. Also, it led to tests using `-race` to fail because the shared
config.klient object that were updating the same fields when initializing
the client.

This commit creates a new `deepCopyConfig` method that allows to create
a deep copy of the config object. This way, each test can have its own
without impacting the others. Now the config has the following lifecycle:
- It is created when a testEnv is created
- Each test uses a child testEnv that inherits the main testEnv's config
- Each feature uses a child testEnv that inherits the test's testEnv's config

This way, a feature inherits all the changes made in BeforeEach functions
while not impacting the other features.